pyproject.toml
[project]
name = "kafka-stock-market"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python-ng>=2.2.3",
    "pandas>=3.0.0",
    "s3fs>=2026.2.0",
]
